requirements.txt
llama-cpp-python>=0.3.0
tokenizers>=0.20.0
huggingface_hub>=0.25.0
duckdb>=1.4.4
matplotlib>=3.8.0
pandas>=2.0.0
tabulate>=0.9.0
tqdm>=4.66.0
